Refactoring: Simplified `PortType`
As the `id` of a `PortType` equals the ordinal, this is exchanged.
Retrieving the `PortType` from a `PortDescriptor`-ID is also moved to
the `PortType`.

An exception for "invalid" IDs is added, this causes the Constructor of
the `PortDescriptor` to throw an `ArtNetException`. This should be
changed with some `ArtNetRuntimeException` to be added at a later state.